+++ Stormpath.SDK/Stormpath.SDK/Linq/AsyncQueryableExtensions.cs
@@ -197,7 +197,7 @@ namespace Stormpath.SDK
         /// <exception cref="ArgumentNullException"><paramref name="source"/> is null.</exception>
         /// <exception cref="OverflowException">The number of elements in <paramref name="source"/> is larger than <see cref="int.MaxValue"/>.</exception>
         /// <exception cref="InvalidOperationException">The underlying provider does not support this operation.</exception>
-        public static async Task<int> CountAsync<TSource>(this IAsyncQueryable<TSource> source, CancellationToken cancellationToken = default(CancellationToken))
+        public static async Task<long> CountAsync<TSource>(this IAsyncQueryable<TSource> source, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (source == null)
                 throw new ArgumentNullException(nameof(source));
